Use `heapless::Vec` in functions `read_many` and `write_many`.
In order to make it possible to use the hardware controlled chip-select,
the `read_many` and `write_many` functions can only contain one
SPI *write* or *transfer* call.
A simple solution to make this possible, is allocating a buffer (or
`heapless::Vec` in this case) in the function itself and concatenate the
register and data there. The biggest downside is that we can no longer
send data of arbitrary length. However, the largest register that one
could read from or write to, is the 64-byte FIFO buffer.
So if we still make this possible (by making the `heapless::Vec` large
enough), this solution should be fine for now.